Split TCPAllocation into tcp_alloc.go
